Move validation utils to validation module
