+++ ethpm/dependencies.py
@@ -1,6 +1,6 @@
 from typing import Dict, List, Tuple
 
-from ethpm.validation import validate_package_name
+from ethpm.validation.package import validate_package_name
 
 
 class Dependencies:
