+++ ethpm/validation/manifest.py
@@ -112,12 +112,6 @@ def validate_manifest_exists(manifest_id: str) -> None:
         )
 
 
-def format_manifest(manifest: Manifest, *, prettify: bool = None) -> str:
-    if prettify:
-        return json.dumps(manifest, sort_keys=True, indent=4)
-    return json.dumps(manifest, sort_keys=True, separators=(",", ":"))
-
-
 def validate_raw_manifest_format(raw_manifest: str) -> None:
     """
     Raise a ValidationError if a manifest ...
