+++ tests/ethpm/test_dependencies.py
@@ -3,7 +3,7 @@ import pytest
 from ethpm import Package
 from ethpm.dependencies import Dependencies
 from ethpm.exceptions import ValidationError
-from ethpm.validation import validate_build_dependency
+from ethpm.validation.package import validate_build_dependency
 
 
 @pytest.fixture
