+++ tests/ethpm/test_manifest_assets_are_valid.py
@@ -4,7 +4,7 @@ import pytest
 
 from ethpm import ASSETS_DIR
 from ethpm.exceptions import InsufficientAssetsError
-from ethpm.utils.manifest_validation import validate_manifest_against_schema
+from ethpm.validation.manifest import validate_manifest_against_schema
 
 SOURCES_GLOB = "**/*.json"
