+++ tests/ethpm/utils/test_manifest_validation_utils.py
@@ -2,14 +2,14 @@ import pytest
 
 from ethpm import V2_PACKAGES_DIR
 from ethpm.exceptions import ValidationError
-from ethpm.utils.manifest_validation import (
+from ethpm.validation.manifest import (
     extract_contract_types_from_deployments,
     validate_manifest_against_schema,
     validate_manifest_deployments,
     validate_manifest_exists,
     validate_raw_manifest_format,
 )
-from ethpm.validation import validate_manifest_version
+from ethpm.validation.package import validate_manifest_version
 
 
 def test_validate_raw_manifest_configuration_validates_strict_manifests(
